% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_shewhart_rules.R
\name{apply_shewhart_rules}
\alias{apply_shewhart_rules}
\title{Apply Shewhart Control Rules}
\usage{
apply_shewhart_rules(data, limits, type)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{limits}{A list containing the lower and upper control limits.}

\item{type}{An integer representing the type of control chart. There are two possible types: \code{1} for discrete data and \code{2} for continuous data.}
}
\value{
A vector of indices corresponding to data points that meet the Shewhart rules.
}
\description{
Identifies points in a control chart that violate Shewhart control rules.
}
\examples{
data_values <- Montgomery2005$Defective
sample_limits <- calculate_limits_type1(Montgomery2005$Defective, Montgomery2005$Sample)
apply_shewhart_rules(data_values, sample_limits)
}
