% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_limits_type1.R
\name{calculate_limits_type1}
\alias{calculate_limits_type1}
\title{Calculate Control Limits (Type 1)}
\usage{
calculate_limits_type1(data, sizes)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{sizes}{A numeric vector of sample sizes.}
}
\value{
A list containing the lower and upper control limits, and the center.
}
\description{
Calculates control limits for type 1 control charts using given data and sizes.
}
\examples{
data("Montgomery2005")
calculate_limits_type1(Montgomery2005$Defective, Montgomery2005$Sample)
}
