% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_limits_type2.R
\name{calculate_limits_type2}
\alias{calculate_limits_type2}
\title{Calculate Control Limits (Type 2)}
\usage{
calculate_limits_type2(data)
}
\arguments{
\item{data}{A numeric vector of data values.}

\item{sizes}{A numeric vector of sample sizes.}
}
\value{
A list containing the lower and upper control limits, and the center.
}
\description{
Calculates control limits for type 2 control charts using given data and sizes.
}
